+++ src/connectors/impls/http/meta.rs
@@ -12,17 +12,17 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-use super::client;
-use super::utils::{FixedBodyReader, RequestId, StreamingBodyReader};
+use super::{
+    client,
+    utils::{FixedBodyReader, RequestId, StreamingBodyReader},
+};
 use crate::connectors::{prelude::*, utils::mime::MimeCodecMap};
 use async_std::channel::{unbounded, Sender};
 use either::Either;
-use http_types::headers::HeaderValues;
-use http_types::Response;
 use http_types::{
-    headers::{self, HeaderValue},
+    headers::{self, HeaderValue, HeaderValues},
     mime::BYTE_STREAM,
-    Method, Mime, Request,
+    Method, Mime, Request, Response,
 };
 use std::str::FromStr;
 use tremor_value::Value;
@@ -43,6 +43,64 @@ pub(crate) struct HttpRequestBuilder {
     codec_overwrite: Option<String>,
 }
 
+#[derive(Clone)]
+pub(crate) struct HeaderValueValue<'v> {
+    finished: bool,
+    idx: usize,
+    v: &'v Value<'v>,
+}
+
+impl<'v> HeaderValueValue<'v> {
+    pub fn new(v: &'v Value<'v>) -> Self {
+        Self {
+            idx: 0,
+            finished: false,
+            v,
+        }
+    }
+}
+
+impl<'v> Iterator for HeaderValueValue<'v> {
+    type Item = HeaderValue;
+
+    fn next(&mut self) -> Option<Self::Item> {
+        if self.finished {
+            None
+        } else if let Some(a) = self.v.as_array() {
+            loop {
+                if a.is_empty() {
+                    self.finished = true;
+                    return None;
+                }
+                let v = a.get(self.idx)?;
+                self.idx += 1;
+                if let Some(v) = v.as_str().and_then(|v| HeaderValue::from_str(v).ok()) {
+                    return Some(v);
+                } else if let Ok(v) = HeaderValue::from_str(&v.encode()) {
+                    return Some(v);
+                }
+            }
+        } else if let Some(v) = self.v.as_str().and_then(|v| HeaderValue::from_str(v).ok()) {
+            self.finished = true;
+            Some(v)
+        } else if let Ok(v) = HeaderValue::from_str(&self.v.encode()) {
+            self.finished = true;
+            Some(v)
+        } else {
+            self.finished = true;
+            None
+        }
+    }
+}
+
+impl<'v> headers::ToHeaderValues for HeaderValueValue<'v> {
+    type Iter = HeaderValueValue<'v>;
+
+    fn to_header_values(&self) -> http_types::Result<Self::Iter> {
+        Ok(self.clone())
+    }
+}
+
 // TODO: do some deduplication with SinkResponse
 impl HttpRequestBuilder {
     pub(super) fn new(
@@ -50,7 +108,6 @@ impl HttpRequestBuilder {
         meta: Option<&Value>,
         codec_map: &MimeCodecMap,
         config: &client::Config,
-        configured_codec: &str,
     ) -> Result<Self> {
         let request_meta = meta.get("request");
         let method = if let Some(method_v) = request_meta.get("method") {
@@ -72,7 +129,6 @@ impl HttpRequestBuilder {
             config.url.clone()
         };
         let mut request = Request::new(method, url.url().clone());
-        let headers = request_meta.get("headers");
 
         // first insert config headers
         for (config_header_name, config_header_values) in &config.headers {
@@ -87,20 +143,11 @@ impl HttpRequestBuilder {
                 }
             }
         }
+        let headers = request_meta.get("headers");
         // build headers
         if let Some(headers) = headers.as_object() {
             for (name, values) in headers {
-                if let Some(header_values) = values.as_array() {
-                    let mut v = Vec::with_capacity(header_values.len());
-                    for value in header_values {
-                        if let Some(header_value) = value.as_str() {
-                            v.push(HeaderValue::from_str(header_value)?);
-                        }
-                    }
-                    request.append_header(name.as_ref(), v.as_slice());
-                } else if let Some(header_value) = values.as_str() {
-                    request.append_header(name.as_ref(), header_value);
-                }
+                request.append_header(name.as_ref(), HeaderValueValue::new(values));
             }
         }
 
@@ -111,29 +158,7 @@ impl HttpRequestBuilder {
 
         let header_content_type = request.content_type();
 
-        let codec_overwrite = header_content_type
-            .as_ref()
-            .and_then(|mime| codec_map.get_codec_name(mime.essence()))
-            // only overwrite the codec if it is different from the configured one
-            .filter(|codec| *codec != configured_codec)
-            .map(ToString::to_string);
-        let codec_content_type = codec_overwrite
-            .as_ref()
-            .and_then(|codec| codec_map.get_mime_type(codec.as_str()))
-            .or_else(|| codec_map.get_mime_type(configured_codec))
-            .and_then(|mime| Mime::from_str(mime).ok());
-
-        // extract content-type and thus possible codec overwrite only from first element
-        // precedence:
-        //  1. from headers meta
-        //  2. from overwritten codec
-        //  3. from configured codec
-        //  4. fall back to application/octet-stream if codec doesn't provide a mime-type
-        let content_type = Some(
-            header_content_type
-                .or(codec_content_type)
-                .unwrap_or(BYTE_STREAM),
-        );
+        let (codec_overwrite, content_type) = consolidate_mime(header_content_type, codec_map);
 
         // set the content type if it is not set yet
         if request.content_type().is_none() {
@@ -235,6 +260,33 @@ impl HttpRequestBuilder {
             None
         }
     }
+}
+
+/// extract content-type and thus possible codec overwrite only from first element
+/// precedence:
+///  1. from headers meta
+///  4. from the `*/*` codec if one was supplied
+///  3. fall back to application/octet-stream if codec doesn't provide a mime-type
+pub(crate) fn consolidate_mime(
+    header_content_type: Option<Mime>,
+    codec_map: &MimeCodecMap,
+) -> (Option<String>, Option<Mime>) {
+    let codec_overwrite = if let Some(header_content_type) = &header_content_type {
+        codec_map.get_codec_name(header_content_type.essence())
+    } else {
+        codec_map.get_codec_name("*/*")
+    }
+    .cloned();
+    let codec_content_type = codec_overwrite
+        .as_ref()
+        .and_then(|codec| codec_map.get_mime_type(codec.as_str()))
+        .and_then(|mime| Mime::from_str(mime).ok());
+    let content_type = Some(
+        header_content_type
+            .or(codec_content_type)
+            .unwrap_or(BYTE_STREAM),
+    );
+    (codec_overwrite, content_type)
 }
 
 /// Extract request metadata
@@ -341,10 +393,8 @@ mod test {
             &Alias::new("flow", "http"),
         )?;
         let config = client::Config::new(&c)?;
-        let configured_codec = "json";
 
-        let mut b =
-            HttpRequestBuilder::new(request_id, meta, &codec_map, &config, configured_codec)?;
+        let mut b = HttpRequestBuilder::new(request_id, meta, &codec_map, &config)?;
 
         let r = b.finalize(&mut s).await?.ok_or("no data")?;
         assert_eq!(